Refactor chat API to use the Cloudflare Agents SDK
This commit refactors the chat functionality to use the Cloudflare
Agents SDK, replacing the direct calls to the AI provider.

A new `ChetAgent` class is introduced to encapsulate the logic for
handling chat requests and streaming responses. The chat API endpoint
now delegates requests to this agent.

The project's dependencies and configuration have been updated to
support the new agent-based architecture.
